+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## v0.0.4 (2026-03-17)
+
+### Breaking
+- default `clientAuthMethod` to `client_secret_basic` if client secret is configured
+
 ## v0.0.3 (2026-03-16)
 
 Refactor cookie handling that unfortunately breaks v0.0.1-2 created cookies. Latest [jose](https://www.npmjs.com/package/jose) uses nodejs builtin [Web Crypto API](https://nodejs.org/docs/latest-v22.x/api/webcrypto.html) ([mdn](https://developer.mozilla.org/en-US/docs/Web/API/Web_Crypto_API)).